fix npe when unloading persistent partitioned topic

+++ pulsar-broker/src/main/java/org/apache/pulsar/broker/service/BrokerService.java
@@ -1682,7 +1682,8 @@ public class BrokerService implements Closeable, ZooKeeperCacheListener<Policies
     public void cleanUnloadedTopicFromCache(NamespaceBundle serviceUnit) {
         for (String topic : topics.keys()) {
             TopicName topicName = TopicName.get(topic);
-            if (serviceUnit.includes(topicName)) {
+            if (getTopicReference(topic).isPresent() && serviceUnit.includes(topicName)) {
+                log.info("[{}][{}] Clean unloaded topic from cache.", serviceUnit.toString(), topic);
                 pulsar.getBrokerService().removeTopicFromCache(topicName.toString(), serviceUnit);
             }
         }
@@ -1709,9 +1710,11 @@ public class BrokerService implements Closeable, ZooKeeperCacheListener<Policies
                     .get(namespaceName);
             if (namespaceMap != null) {
                 ConcurrentOpenHashMap<String, Topic> bundleMap = namespaceMap.get(bundleName);
-                bundleMap.remove(topic);
-                if (bundleMap.isEmpty()) {
-                    namespaceMap.remove(bundleName);
+                if (bundleMap != null) {
+                    bundleMap.remove(topic);
+                    if (bundleMap.isEmpty()) {
+                        namespaceMap.remove(bundleName);
+                    }
                 }
 
                 if (namespaceMap.isEmpty()) {
